Alter how we detect username from a social authenticator (#622)
In playing around with social auths, some of them don't seem to have a
way to remap attributes to what we need (like username) but there is a
built in function that gets an "ID_KEY". This change will attempt to get
the ID_KEY for an authenticator and then fallback to our expected
`username`.